clang: we can now build all of userland and boot
usbd dies with a suicide, as it seems to try to use floating point
in a note handler, oops. I bet there is a memcpy and clang generates
code to use xmm.

I remove the call to build libz/regression.json because clang got upset
with it.

There are many warnings, still, but at least we can build. I am going to get the warnings
fixed tonight/tomorrow and then add clang to the travis test matrix.

+++ sys/src/libstdio/vfprintf.c
@@ -297,7 +297,7 @@ ocvt_fixed(FILE *f, va_list *args, int flags, int width, int precision,
 	char digits[128];	/* no reasonable machine will ever overflow this */
 	char *sign;
 	char *dp;
-	int32_t snum;
+	int64_t snum;
 	unsigned long num;
 	int nout, npad, nlzero;
